[REFACTOR] 댓글 조회 로직 분리 #86
